Moved handlers into attachment card
